Correction des problèmes de timings dans Hml pour les calculs de vitesse : _ passage en long double _ utilisation de l'info TimeCode des faulhaber _ réduction période du CAN
git-svn-id: svn://team-ard.com/ARP/trunk@1062 c9b90f90-cbc1-11df-9d65-af01d29d43da

+++ arp_core/src/math/EstimatedPose2D.cpp
@@ -44,12 +44,12 @@ Eigen::Matrix<double,3,3>& EstimatedPose2D::covRef()
     return covariance;
 }
 
-double EstimatedPose2D::date() const
+long double EstimatedPose2D::date() const
 {
     return estimationDate;
 }
 
-double& EstimatedPose2D::dateRef()
+long double& EstimatedPose2D::dateRef()
 {
     return estimationDate;
 }
@@ -59,7 +59,7 @@ void EstimatedPose2D::cov(Eigen::Matrix<double,3,3> _cov)
     covariance = _cov;
 }
 
-void EstimatedPose2D::date(double _date)
+void EstimatedPose2D::date(long double _date)
 {
     estimationDate = _date;
 }
